requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
amqp==5.1.1
    # via
    #   -r requirements/base.txt
    #   kombu
aniso8601==9.0.1
    # via
    #   -r requirements/base.txt
    #   tincan
apache-libcloud==3.7.0
    # via -r requirements/base.txt
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
backports-zoneinfo[tzdata]==0.2.1
    # via
    #   -r requirements/base.txt
    #   celery
    #   kombu
billiard==4.1.0
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.3.0
    # via
    #   -r requirements/base.txt
    #   edx-celeryutils
    #   event-tracking
certifi==2023.5.7
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
click-didyoumean==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.2.0
    # via
    #   -r requirements/base.txt
    #   celery
code-annotations==1.3.0
    # via
    #   -r requirements/base.txt
    #   -r requirements/test.in
    #   edx-toggles
coverage[toml]==7.2.7
    # via pytest-cov
cryptography==41.0.1
    # via
    #   -r requirements/base.txt
    #   djfernet
ddt==1.6.0
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-config-models
    #   django-crum
    #   django-model-utils
    #   djangorestframework
    #   edx-celeryutils
    #   edx-django-utils
    #   edx-toggles
    #   event-tracking
    #   jsonfield
django-config-models==2.3.0
    # via -r requirements/base.txt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-toggles
django-model-utils==4.3.1
    # via
    #   -r requirements/base.txt
    #   edx-celeryutils
django-waffle==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-toggles
djangorestframework==3.14.0
    # via
    #   -r requirements/base.txt
    #   django-config-models
djfernet==0.8.1
    # via -r requirements/base.txt
edx-celeryutils==1.2.2
    # via -r requirements/base.txt
edx-django-utils==5.5.0
    # via
    #   -r requirements/base.txt
    #   django-config-models
    #   edx-toggles
    #   event-tracking
edx-toggles==5.0.0
    # via -r requirements/base.txt
event-tracking==2.1.0
    # via -r requirements/base.txt
exceptiongroup==1.1.1
    # via pytest
factory-boy==3.2.1
    # via -r requirements/test.in
faker==18.10.1
    # via factory-boy
fasteners==0.18
    # via -r requirements/base.txt
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via -r requirements/base.txt
jinja2==3.1.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
jsonfield==3.1.0
    # via
    #   -r requirements/base.txt
    #   edx-celeryutils
kombu==5.3.0
    # via
    #   -r requirements/base.txt
    #   celery
markupsafe==2.1.3
    # via
    #   -r requirements/base.txt
    #   jinja2
mock==5.0.2
    # via -r requirements/test.in
newrelic==8.8.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==23.1
    # via pytest
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.38
    # via
    #   -r requirements/base.txt
    #   click-repl
psutil==5.9.5
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   event-tracking
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==7.3.1
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   celery
    #   faker
python-slugify==8.0.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
    #   event-tracking
    #   tincan
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
requests==2.31.0
    # via
    #   -r requirements/base.txt
    #   apache-libcloud
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   click-repl
    #   event-tracking
    #   isodate
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.1.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
text-unidecode==1.3
    # via
    #   -r requirements/base.txt
    #   python-slugify
tincan==1.0.0
    # via -r requirements/base.txt
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.6.3
    # via
    #   -r requirements/base.txt
    #   asgiref
    #   kombu
tzdata==2023.3
    # via
    #   -r requirements/base.txt
    #   backports-zoneinfo
    #   celery
urllib3==2.0.3
    # via
    #   -r requirements/base.txt
    #   requests
vine==5.0.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit
